Correctly display created option in Select
Before this change, we made some internal formatting in `Select` to
allow us display the created values when `allowCreate` and `multi` were set.
This did not affect `Select` components with only one value, so the
options created in them did not correctly appear in the field, even if
the value was correctly pass in the component.

This change ensures the formatting function is called regardless of the
value of `multi`.