Splitting out Plotly from main bundle.
This change is creating two new components, `Plot` and `AsyncPlot`. The
first is bundling the bootstrapping code to generate a `Plot` component
from a custom plotly configuration, the latter is loading that one
asynchronously.

Using the `AsyncPlot` component in all components lead to splitting out
the complete `plotly` dependendency, resulting in a large size reduction
of the entry bundle.