Index mapping changes (#6039)
* Add "match_mapping_type" to the dynamic mapping for "gl2_*" fields

The previous mapping configuration forced all "gl2_*" fields which have
not been mapped explicitly to string/keyword. That means we couldn't
dynamically add new "gl2_*" fields which data types of boolean, long,
date or others.

With this commit we add "match_mapping_type" with value "string" to make
sure that all "gl2_*" string fields will be indexed as keyword instead
of text and others will be mapped automatically by Elasticsearch.

This allows us to index date fields and others with the correct type
without having to create an explicit mapping.

* Rename "triggered_tasks" mapping to "triggered_jobs" in events template

This has been changed for consistency. Our scheduler is using jobs and
not tasks.