Do not refresh search if query input is blurred, but query is unchanged. (#6045)
* Do not refresh search if query input is blurred, but query is unchanged.

Before this change, when leaving the query input without changing the
query (e.g. when copying parts of the query to the clipboard), the
search was refreshed nonetheless.

This change is preventing this. Whenever the query input looses focus,
it is checked for changes and the `onChange` callback is executed only
if there was a change. The `onBlur` callback is triggered regardless of
that.

Fixes #6038.

* Use short-hand editor variable.

* Extract placeholder management functions.

* Ensure that `lastValue` is updated when execution is triggered.