Extracting `onZoom` function, using `tozero` for y axis's `rangemode` (#6048)
* Extracting `onZoom` function, using min/max of y values for y axis range

This change does the following:

  * Extract `onZoom` function from component for easier testing. It can
now also be passed as prop, defaulting to the current implementation for
backwards compatibility
  * Setting `rangemode` for y axis to `tozero`. It sets the range
automatically according to the minimum/maximums of the values passed. If
minimum is greater than 0, the lower range of the chart will be set to 0
to ensure that charts with all positive values are aligned

Fixes #6037.

* Removing explicit call to refresh search from `onZoom` handler.

Before this change, refreshing the search was required to be done
explicitly. Now that the search is refreshed automatically upon
structural changes, this is not required and leads to a potentially
wasteful double refresh.

* Removing now unused import.