Splitting out fat dependencies from main bundle. (#6052)
* Loading `QueryInput` asynchronously.

This commit is creating a `AsyncQueryInput` component and uses it in
all places where `QueryInput` was used before. This component loads the
`QueryInput` asynchronously, leading to it and `react-ace-builds` being
splitted out, reducing entry bundle size.

* Splitting out Plotly from main bundle.

This change is creating two new components, `Plot` and `AsyncPlot`. The
first is bundling the bootstrapping code to generate a `Plot` component
from a custom plotly configuration, the latter is loading that one
asynchronously.

Using the `AsyncPlot` component in all components lead to splitting out
the complete `plotly` dependendency, resulting in a large size reduction
of the entry bundle.

* Adding react-overlays flow stub.

* Fixing linter hints.