Mocking FetchProvider globally for tests.
When implicit/transitive dependencies of a SUT require stores that are
not mocked and those stores do remote calls during initialization,
promises are instantiated which are rejected asynchronously and -
depending on the runtime of all tests - may or may not be rejected in
the background. This leads to build failures in some cases (the
probability seems to be higher for concurrent build/test runs) which are
not even assignable to a specific test (case).

This change is mocking `FetchProvider` globally for jest, so no remote
calls can be done using our fetch API. This is a precaution against the
scenario described above and also a good indication for anyone writing a
test that it is not properly mocked (as remote calls should never happen
in test cases).