Add grand total to PaginatedResponse
Before this change, using PaginatedResponse with a query returned a
total of elements matching the query. The response had no indication of
how many items of that entity were there, something problematic for
the frontend, since we want to render a different messages for both
situations.

This commit adds a `grandTotal` field to `PaginatedList` and adds it
into the response if the field is set, making it easier for the frontend
to detect when there are no available items from certain entity.