Improvements in Event Definitions page (#6091)
* Remove unused execute method in EventDefinitionsStore

This resource was meant for testing the execution before the scheduler
was done.

* Add pagination to EventDefinitionsStore

* Add pagination to Event Definition list

* Add filter controls to Event Definition list

* Add grand total to PaginatedResponse

Before this change, using PaginatedResponse with a query returned a
total of elements matching the query. The response had no indication of
how many items of that entity were there, something problematic for
the frontend, since we want to render a different messages for both
situations.

This commit adds a `grandTotal` field to `PaginatedList` and adds it
into the response if the field is set, making it easier for the frontend
to detect when there are no available items from certain entity.

* Update check of empty list of event definitions

* Remove test page size

* Use single element for event definition actions

* Get event definition name from plugin

* Add more information to Event Definition list

The change includes scheduling information and notifications
information.