Apply custom highlighting only if field value matches configured value exactly. (#6098)
Before this change, the configured value of a highlighting rule needed
to be a substring of the field's value only, to make it become
highlighted. This is counter-intuitive, especially for numeric field
values.

This change is changing this so the field value needs to match the
configured rule's value exactly.

Fixes Graylog2/graylog-plugin-enterprise#1118.