Improve shutdown handling of processing status recorder
Don't try to persist the processing statusldata when the server is
shutting down. This avoids sporadic MongoDB driver exceptions when the
service tries to persist the data while the DB driver is shutting down.

The service now also subscribes to the event bus to get the shutdown
signal as early as possible.