Fix rendering of events in single message view. (#6156)
Before this change, rendering a message generated by an event using the
`ShowMessagePage` component, e.g. using the `/messages/{index}/{id}`
endpoint failed, due to a missing source input id field in the event
message.

This change is fixing this and a couple of other issues with rendering
event messages.

Fixes #6142.