Fix "Routed into streams" element on show message page
The problem is that the MessageResource returns a ResultMessage instead
of a ResultMessageSummary. Due to that, we serialize the regular Message
object instead of only the fields. This exposes internals of the Message
object and also breaks the "Routed into streams" element on the show
message page because the UI elements don't expect the Message structure.

This change is basically backwards incompatible change for the get
message API because the response gets changed.