Fix closing of widget filter menu.
Before this change, closing the widget filter menu popover resulted in
it immediately showing again and logging an error to the console. This
was the result of not passing a function for `onChange` to the
`QueryInput` component that returns a Promise.

This change is fixing that by returning a promise from the `onChange`
handler.