Set type default config when type change
Define default configurations for Event Processor Configs and Field Value
Providers in plugin interface and set those values when the type change.
This allows us to control when the default types should be applied, and
do not force plugins to propagate the default configuration in
`componentDidMount`. This is not mandatory, and plugins can set the
default configuration otherwise if it makes sense to do so.