Improve handling of default configurations in Event Definitions and Notifications (#6186)
* Propagate default legacy callback configuration

When selecting the type of legacy alarm callback, set the default type
configuration as the current configuration. This way all default values
will be persisted and used as normal configuration values.

Fixes #6185

* Set notification default config when type changes

Do not use `componentDidMount` on plugins to propagate the default
configuration, define it in the plugin information instead and set it
when the type changes.

* Set type default config when type change

Define default configurations for Event Processor Configs and Field Value
Providers in plugin interface and set those values when the type change.
This allows us to control when the default types should be applied, and
do not force plugins to propagate the default configuration in
`componentDidMount`. This is not mandatory, and plugins can set the
default configuration otherwise if it makes sense to do so.