Improve job scheduler responsiveness (#6224)
The loop used to sleep for a second when there were no free worker pool
slots or when there were no runnable triggers. This was done to avoid
busy spinning.

This was an issue when jobs need to catch up a big backlog and their
execution time is very short.

The job execution engine is now publishing an event when a job execution
has been completed. The scheduler service is subscribing to these events
and interrupts the sleep when receiving such an event to immediately
check for another job to execute.

We are now also waiting for the server to reach RUNNING status before we
start the scheduling loop.

Closes #1107

* Fix InterruptibleSleeper implementation

Calling "release()" on a semaphore will *always* increase available
permits and does not cap permits once the initial permit count is
reached.