Removing Dashboard state from Views.
The existing dashboard state in the view model is not used anymore since
the dashboard tab was removed. This change is now removing the residues
of this, in order to be able to replace it with a better suited data
structure to support the presentation mode of future dashboards better.