Implement permission check for events search
The events search code is now checking if the user that executes the
search is allowed to see events based on the "source_streams" field.

Example:

If an event got created from a message in stream A, the event has stream
A in its "source_stream" field. If a user has read permissions on stream
A, the event is visible to the user in the events search.

Closes #6238