Switch node liveness detection from nodes status to updated_at field
Previously we use the node status collection to get all active nodes.
Since that system is currently pretty agressive removing nodes from the
database, this change switches the liveness detection of nodes over to
using the "updated_at" timestamp of the processing status data.

The default threshold is set to one minute. That means if a node doesn't
update its processing status for one minute (which usually happens every
second by default), its data is not considered in the check.

Refs #6228