Prevent closing of UDP channel in input for messages with bogus GELF header.
Before this change, whenever a UDP message came in with a 0 or 1 byte
header (smaller than the 2 byte GELF header), the thread handling this
message throws an exception in `GELFMessage#getGELFType` which is caught
in `EnvelopeMessageHandler#exceptionCaught`. In there, the channel is
closed. This is correct for a TCP-based input, as the corresponding
TCP-connection would be closed, as each channel corresponds to a TCP
connection. This is not the case for UDP pipelines. Closing a channel of
a UDP pipeline leads to the handler/thread losing the overall connection
to the socket multiplexing. If this happens for a number of times
equaling the number of working threads, the whole input stops processing
any UDP messages as there is no thread left which is able to receive
channel events.

This change adds `EnvelopeMessageAggregationHandler#exceptionCaught`
which is only logging the exception and increasing the counter for
invalid GELF chunks, without propagating the exception to other handlers
(preventing the channel closure).

A remaining issue is that for UDP messages with no/invalid GELF headers,
the `EnvelopeMessageAggregationHandler#channelRead0` is called four
times, for correct messages only once. The first three times it is
called with a single byte buffer, the last time with a buffer containing
the complete message. It is unclear to me why this happens and leads to
the exception being logged four times, but this seems to be still way
better than causing situations where the complete input ceases to work.
Any insight how to fix this is welcome.

Fixes #5701.