Fix retrieval of registered plugin routes.
Before this change, retrieval of plugin routes using
`Routes.pluginRoute('...')` had the following issue:

  1. Plugin routes were retrieved once and then stored in
`window.pluginRoutes`
  2. Every consumer (core or any plugin) needs to import `Routes.jsx` to
be able to retrieve a plugin route
  3. When the file is imported from a different scope (e.g. core vs
plugin, or plugin1 vs. plugin2), the file is evaluated again, performing
1. again.

This means that the contents of `window.pluginRoutes` could be different
depending on when an where the module is imported. This could lead to a
scenario of:

  * Plugin A requires a plugin route of Plugin B at runtime.
  * If Plugin B was not loaded before Plugin A, `window.pluginRoutes` does not
contain Plugin B's routes if it was only imported from Plugin A (earlier),
although they are available at runtime (the actual route retrieval)

This change is fixing this behavior by not caching the plugin routes,
but iterating over all `route` exports in the `PluginStore` and
retrieving the matching route from there, always returning accurate
results.