Implement new getPackageName function for versionConstraints
Prior to this change, the correlation config did not add its enterprise
plugin constraint to the constraints list.

This change adds a new function to the ContentPackable constraint:
getPackageName.

This is necessary since the package name of CorrelationConfig is
graylog.plugins.events but we need graylog.plugin.enterprise since
it is the name of the constraint.