Place Action Menus bottom to target
Prior to this change, when the action menu (either value or field) was
rendered, it was drawn right to the target element. This led to the
problem that it could be rendered outside of the viewport.

This change will render the menu below the target element.
When the menu is still rendered outside the scrollbar grows and you
can scroll to the desired action.

Fixes #6282