Wait for `ViewActions.load` in promise chain and still return view. (#6352) (#6361)
Before this change, calling the `ViewActions.load` action from
`ViewDeserializer` was not returning the promise of it, so it was not
waited upon to resolve. This lead to an issue, where loading the view
took longer than toggling the `loaded` status flag in `ShowViewPage`,
leading to rendering components which relied on the view being loaded
already.

This change is now returning the promise received from
`ViewActions.load`, so it is waited on.

Fixes #6346.