Allowing to enable (experimental) frontend features for development. (#6367)
This change is adding a simple way to enable frontend features for
development which are disabled per default. The change is two-fold:

  * Evaluating a `FEATURES` environment variable which contains a
comma-delimited string of feature names to enable
  * Adding a `AppConfig.isFeatureEnabled` function, which returns a
boolean indicating if the given feature is enabled or not. This can be
used in the code to conditionally render parts of the UI.

Notes:

The content of the `FEATURES` can contain multiple features separated by
commas. Additionally, individual feature flags are compared
case-insensitively, so `FEATURES=Time_Travel` will enable the
`time_travel` feature as well.

Example:

If a feature `time_travel` is developed but should not be enabled by
default, developers can do a `FEATURES=time_travel yarn run start` to
enable it for their development environment. In the code, something like
this can be done:

```
<div>
  {AppConfig.isFeatureEnabled('time_travel') && <TimeTravel />}
  ...
  {AppConfig.isFeatureEnabled('time_travel') || <OldBoringStuff />}
</div>
```