Removing the system HTTP client provider.
The system HTTP client provider was used to consume a special http
client that is completely bypassing the proxy configuration. This is not
required anymore, as we have the `http_non_proxy_hosts` pattern which
allows blacklisting of hosts/host patterns circumventing the proxy
configuration.

This change requires advice in the documentation for plugin authors, as
the system HTTP client provider might have been used in 3rd party
plugins.