Fix KafkaJournal metrics on inactive nodes
If nothing was ever commited to a journal,
the uncommitedMessages gauge always reported 9223372036854775807 (Long.MAX_VALUE)

This happens because committedOffset is initialized as DEFAULT_COMMITTED_OFFSET (Long.MIN_VALUE)
(substracted by 1 underflows the long to Long.MAX_VALUE)

This breaks the idle node detection feature for the processing status,
which would exclude entries that have no messages in the journal.

Steal the code from the ThrottleStateUpdaterThread, which already deals with this edge case.

Fixes #6415