Do not try to add nonexisting subcharts to stacked chart. (#6432)
Before this change, stacked charts on the current search page could
reference a nonexisting subchart (see #6355 for an example), leading to
a nonfunctioning search page where the user is not able to remove the
faulty chart to restore it.

This change is adding subcharts conditionally, skipping them if the
given reference cannot be resolved. Therefore, even stacked chart
definitions which reference nonexisting subcharts can be displayed
(partially) correct.

Fixes #6355.