Fixing event leak prevention in `Widget` component. (#6433)
In #6431 a event propagation barrier was implemented that was supposed
to prevent mouse events from leaking the `Widget` component. This was
done in a too radical way, also preventing valid mouse events (when
dragging a widget in dashboard edit mode) from reaching the grid
container.

This change is now preventing mouse events from leaking the edit/show
config modals only.