Do not create compound types if types differ by properties.
Before this change, compound types were created also for these types:

  * `text` -> `string{properties=[full-text-search]}`
  * `keyword` -> `string{properties=[enumerable]}`

This is incorrect, as these are the same (mapped) types, but vary only
on capabilities.

This change is now merging these two types into a single non-compound
type, with a property set that is the intersection of the given
properties of all merged types.

Fixes #6313.