Add custom EventDto DateTime deserializer
EventDto timestamps can come in three different formats:
 - Already converted to ISO8601
 - Our custom ES_DATE_FORMAT
 - Date object from MongoDB

An earlier approach to solve this used an Converter which
the Problem that it used String values for its input.
This meant that the Date objects from MongoDB were coerced to Strings,
which were presented in a human readable format, which could not be
converted.

Implement a Deserializer which works on JsonParser and can
be used to convert all three formats.