Add custom EventDto DateTime deserializer (#6468)
* Add custom EventDto DateTime deserializer

EventDto timestamps can come in three different formats:
 - Already converted to ISO8601
 - Our custom ES_DATE_FORMAT
 - Date object from MongoDB

An earlier approach to solve this used an Converter which
the Problem that it used String values for its input.
This meant that the Date objects from MongoDB were coerced to Strings,
which were presented in a human readable format, which could not be
converted.

Implement a Deserializer which works on JsonParser and can
be used to convert all three formats.

* Ignore "_id" in EventDto to replace some hardcoded behavior

The "_id" field gets added by our legacy search system and is the same
as the "id" field. Previously we had some code to manually remove the
field.

Also start a test file for EventDto.

Fixes Graylog2/graylog-plugin-enterprise#1231

(cherry picked from commit 6d07c88799188bcce8bc3ab66adc37d71ac95986)