Add stringListValue() support to LookupTables
Since our current use cases for listValues only need String objects,
and we don't want to complicate the API unneccessarily, don't
implement this as a List of Objects.

The idea is that each DataAdapter should decide which values
make most sense to be in the List, and convert them accordingly.