Render `null` for message field if its value is `undefined`.
Before this change, the `PossiblyHighlight` component which is used to
determine if given highlighting rules apply to a message field and its
value and generates highlighting areas for the given field value in this
case returns `undefined`. This violates a React invariant limiting the
result of a component to `React.Node` or `null`.

This change is slightly changing the logic of the affected component and
returns `null` in this case.

Fixes #6511.