Notification test button improvements (#6522)
* Keep test notification message on screen

Before this change, the result of testing a notification was displayed
in a toaster message and gone in a few seconds. This change makes the
message to be displayed below the test button until one of the input
changes.

* Add button to test notification from list

This change allows users to test notifications from within the
notification list.

(cherry picked from commit a56f3a1921b89156f493da93bd2a13fb4ef259d9)