Add ViewTransformer
Prior this change we could not transform a serach view
into a dashboard view. But the user wants to be able
to take a search view and convert it to a dashboard view.

This change will add a new function which takes a search
view and returns a new dashboard view.