Do not render empty string for `0` value in number visualization.
Before this change, the number visualization of the aggregation builder
rendered an empty string if the actual value was `0`, related to a
type coercion in a conditional meant to catch `null`/`undefined` values.

This change is making that conditional stricted. In addition, it also
returns `N/A` instead of `''` for `null`/`undefined` values.