Use useComponent for use of styled
Prior to this change, the function styled was regenerated
with every rendering in the functional components. But
this was not necessary. A update of the function was only
needed if the prop active did change.

With the use of Refs this got even more problamtic since the
styled function was returning a component which was used
by a forwaredRef. The Rerender of the component was then
producing a new component and the ref was stale.

This change will wrap the styped function in a useCallback
to ensure the returned component does not get removed
and the ref does not get stale.