Improving `UserPreferencesButton` tests usage of mocks and snapshots.
Before this change, the `UserPreferencesButton` test was snapshotting
the instance returned from testing-library directly, including a lot of
internal structure in the code which is likely to change and therefore
break the test in future versions of the library. This change is fixing
this. Additionally, it removes a lot of unnecessary boiler plate.