Fixing root route in `AppRouter`.
In #6602 the `AppRouter`'s route structure was changed to allow
supplying routes from plugins with an explicitly defined `null` value
for the `parentComponent`. This enforces hoisting of these routes to the
top level of the router so they are not contained within the `App` or
other parent components.

Unfortunately the path of the global `AppErrorBoundary` was left in
place, erroneously meant to fix a unrelated route matching issue,
leading to a ` ` or `/` route to go nowhere.

This change is removing the path specification so ` ` and `/` routes are
correctly routed to the `StartPage` component again.