Set isNew flag to false when a new view is loaded
Prior to this change, when a new view was loaded the ViewStore
could still have the isNew flag set to true.

This change will add a listener to the ViewManagementStore.
If the get action was completed then we have a not so new View
loaded.

Also remove unused setNew action.