Fix adding more than one percentile aggregation
Prior to this change, the user was unable to create more than one
percentile metric for same field in aggregation. Since the aggregation
series name was matching on aggregation name and field.

This change will add the percentile to the aggregation name. That way
the user can add more then one percentile aggergation.

Fixes: #6597