Check if passed View has a search
Prior to this change, after a reload of the transformed dashboard
view, the state containing the view was still stored somewhere in the
browser in JSON format (JS literal object).
When rendering the NewDashboardPage with that object there was no
search only the search id, which lead to a error.

This change will check if the passed view contains a search (and not
just the ID of a search).

Thanks to @linuspahl for finding this!