Add instance check for flow
Prior to this change, flow was raising concerns that viewTransformer
was getting a ViewJson passed even though the if guared that by checking
if a search attribute is part of the searchView object (which is not
part of the ViewJson type).

Since we cannot use instanceof (which has troubles with our plugins)
still need the searchView.search check.

This change will add instanceof to the guard to sooth flow without
a FlowFixMe.