Remove instanceof since it does not work
Prior to this change, I introduced instanceof as a guard for flow.
But as I mentioned in the git commit this will not work.
For some reason I do not understand I did not trust my own
words and added the instanceof anyway.

This change will remove instanceof and add a FlowFixMe.