Deduplicating query/search type errors before returning them.
Before this change, errors returned from the execution of a `Search` on
a query or search type level were returned as is from ES to the API
consumer, leading to repeating each error for every index it occurs on.

This change is deduplicating errors by only returning distinct
descriptions to the caller.