Add CopyToDashboardForm Component
To copy a widget to a dashboard the user needs to
select the desired dashboard.

This component will render a Dashboard List where
the user can search and select the the desired dashboard
from.

The Component will be rendered as a modal and displayed
when the user clicks on the widget action
"Copy To Dashboard".