Render expression recursively
So far we supposed an aggregation would only have a single condition and
to simplify the tree coming from the server, we flattened it and use
that to render the form.

This change renders the whole expression tree recursively, making it
possible to render the forms for more than one condition. It still
doesn't allow to add, edit or delete expressions from the UI, that will
be done separately.