SessionsResource accepts plain JSON, token creation delegated to factory
The SessionsResource will now accept a JsonNode. Creation of an
authentication token is delegated to a factory so that a plugin can
create its own token classes for the JSON payload.

Authentication tokens are made aware of the audit actor, so that the
sessions creator can emit the appropriate audit event in case a user
was unable to authenticate (in that case we might not have a username
if the username is not part of the credentials in a plugin's custom
token implementation)