Use widget timerange for time bucketing auto interval if present.
Before this change, only the query's time range was considered for the
auto interval of time bucketing. This could lead to a widget showing a
very large time range with a very small bucket size because the auto
interval would use the query's time range which defaults to 300 seconds.

This change is now optionally using the widget/pivot's time range if
present, the query's timerange otherwise.