Tracking overridden query/timerange in `GlobalOverride` object.
Before, overriding the query string/timerange of a Query was applied
directly to the instance when parsing the search execution state, making
it invisible that a change from the global override has happened.
In some cases we want to determine if a time range was set from the
query or from the global override. Therefore this commit introduces an
internal `GlobalOverride` object which tracks optional overrides. Also
it implements an `effectiveTimerange` method for the `Query` object
which allows a search type to determine if the time range from the
global override, from itself or from the query (in that order of
precedence) should be used.