Improve customizability of the authentication system (#6661)
* Extract code for session creation from the SessionsResource

The code for creating a session based on a username/password pair was
located in the SessionsResource. If a plugin wanted to create a
session based on other credentials, this code would have to be
duplicated.

Now there is a SessionCreator which can be used for this task.

* Add unit test for SessionCreator

* Improve javadoc

* Fix license headers

We use "wrong" license headers, but that's how it is.

* Don't catch exception which never seems to be thrown

There is an UnknownSessionException thrown deep down in the shiro code
when a session ID is given but it is unknown. However it was caught
further up in the stack in all the scenarios I tested. So we don't
handle it explicitly to make the code cleaner.

* SessionsResource accepts plain JSON, token creation delegated to factory

The SessionsResource will now accept a JsonNode. Creation of an
authentication token is delegated to a factory so that a plugin can
create its own token classes for the JSON payload.

Authentication tokens are made aware of the audit actor, so that the
sessions creator can emit the appropriate audit event in case a user
was unable to authenticate (in that case we might not have a username
if the username is not part of the credentials in a plugin's custom
token implementation)

* No need for casting

* Allow disabling of authentication realms by configuration

A list of activated realms can be provided in the configuration file.
If this list is configured, only the realms listed will be used for
authentication. The non-activated realms will also not be visible on the
frontend.

* Add license headers

* Move binding of token factory to pre-existing module; remove now empty module

* Add username to session creation response

* Revert "Allow disabling of authentication realms by configuration"

This reverts commit 3077cdf0

* Implement much simpler approach to deactivating authentication realms

We focus on the builtin realms only and therefore can simply skip
adding the respective binding for a realm if deactivated.
This is much simpler but should be sufficient for our use case.

* No wildcard imports.