Use widget timerange for time bucketing auto interval if present. (#6682)
* Use widget timerange for time bucketing auto interval if present.

Before this change, only the query's time range was considered for the
auto interval of time bucketing. This could lead to a widget showing a
very large time range with a very small bucket size because the auto
interval would use the query's time range which defaults to 300 seconds.

This change is now optionally using the widget/pivot's time range if
present, the query's timerange otherwise.

* Adding license header.

* Tracking overridden query/timerange in `GlobalOverride` object.

Before, overriding the query string/timerange of a Query was applied
directly to the instance when parsing the search execution state, making
it invisible that a change from the global override has happened.
In some cases we want to determine if a time range was set from the
query or from the global override. Therefore this commit introduces an
internal `GlobalOverride` object which tracks optional overrides. Also
it implements an `effectiveTimerange` method for the `Query` object
which allows a search type to determine if the time range from the
global override, from itself or from the query (in that order of
precedence) should be used.

* Considering time range from global override when adding effective time range to results.

* Mediating license escalation.