Hide root user when the RootAccountRealm is deactivated (#6688)
Even when "root-user" was listed under
"deactivated_builtin_authentication_providers" in the configuration
file, the root user would still be shown on the frontend and used at
some places in the code. Having the root-user linger even though it
can't really be used is confusing.

The root user will now be hidden in the frontend and it has been made
clearer in the API that the root user might not exist (by making the
return value optional).

Setting the root user's password in the config is not required anymore
when the user is deactivated.

* Use fixed "local:admin" user ID for legacy email callback migration

The user ID of the local admin user is always "local:admin" regardless
of the configured username in the configuration. We simply use this
string to let the migration run, even if the local admin user is
disabled. This is safer than throwing an exception.