Include streams of search type when returning used streams of search.
Before this change, `Search#usedStreamIds` only considered the
referenced streams of `Query`'s. This led to an omission of permission
checks for the search type's stream ids.

This change is now returning the stream ids referenced in the search
types of a query as well for `Search#usedStreamIds`.

In addition, it now returns an empty set instead of throwing an
exception when performed on a search with an empty set of queries (a
pathological case, but returning an empty set seems to be the correct
way to handle it).