Prevent stale objects in stream cache
Previously we were storing the stream itself in the nameToStream
MultiMap but were using only the id of the stream as the comparator for
the underlying set. This would lead to stale stream objects in that
cache if the cache was updated with a changed stream which was already
cached.

By using a real cache to load the streams by name on demand and
invalidating if a stream changes, we impose a penalty on first lookup by
name after a change but avoid the stale items.

fixes #4954