Migrate non-Reflux stores to Reflux
Replace old Typescript stores fetching data to Reflux stores. This does
not add actions or any advanced Reflux feature, but leaves the stores
simple to avoid creating a larger changeset.

Using Reflux ensures stores' methods are bound to the object and will
execute in the right execution context, avoiding issues as the one
described in #6704.

Fixes #6704