Changing views integration test to test dashboards.
The `CreateNewView` test is becoming obsolete due to the functionality
changing from generic views to specialized searches/dashboards.
Therefore, and because the test was doing too much (rendering a full
view with widgets while ending the test before the rendering was done),
the test is changed to test general dashboard creation functionality.