Disable ace's indentation for `QueryInput`.
This PR disables ace's integrated indentation feature, which captures
the tab key to indent text. As this does not make much sense for a
single line input, but does not allow the user to use the keyboard to
jump out of this input and the focus is trapped, this feature is
disabled.