Adding `DerivedTimeRange` class to wrap time ranges for search types.
This class is supposed to wrap `TimeRange` instances in some contexts
(currently search types) where the `TimeRange` could also be an
implementation of the `OffsetRange` time range, which requires
additional dependencies (`Query`/`SearchType`) to calculate the
effective time range.

The class is wrapping the actual `TimeRange` and the only way to
retrieve it from an instance is to call `effectiveTimeRange` and pass in
the required dependencies for its calculation.

Whenever the `OffsetRange` is supposed to be used, it must be wrapped in
a `DerivedTimeRange` instance, as it only contains the parameters to
reference and offset an existing time range, but is not able to
calculate it on its own and therefore throws an ISE when
`getFrom`/`getTo` is called.