Fix COMMONAPACHELOG grok pattern.
The COMMONAPACHELOG grok pattern uses the original date string found in
the source to propagate the "timestamp" message field. When the
extractor tries to use this value to set the field on the message, it
will be ignored, as it is not a proper date.

We adjust the pattern so that it creates a date object which the message
accepts as a valid value for the "timestamp" field.

Fixes #1647