Switch Elasticsearch IT tests to use testcontainers (#5391)
* Add a ElasticsearchBaseTest class based on the testcontainers project

This is supposed to be a replacement for the nosqlunit based
ElasticsearchBase class. It's using the testcontainers library to
provide Elasticsearch instances for integration test executions.

* Switch all existing ES integration tests over to ElasticsearchBaseTest

This makes the nosqlunit based ElasticsearchBase obsolete.

* Enable INFO log level for "org.graylog" in the test log4j2 config

* Delete unused ElasticsearchBase test class

* Remove old elasticsearch integration test maven configuration

- Remove unused "it.es.*" properties
- Only run ES integration tests when "it.es" property is enabled
- Remove elasticsearch-maven-plugin dependency and config

* Switch log statements to DEBUG in ElasticsearchInstance

* Limit elasticsearch instances to 512 MB memory

This should be enough to run tests.

* Remove now unused nosqlunit-elasticsearch-http dependency

* Reuse Elasticsearch containers

Instead of creating a new container for every test, we reuse one for all
of them. Indices and Templates are cleared between tests.
Before this change every affected test needed about 10s to spin up the
container and ES instance. Now most of them need well under 1s.

* Remove unnecessary cleanup code in ES integration tests

* Add settings to ES container

discovery.type=single-node should avoid clashes with other es containers
running in the same network. there should be none right now, but we want
to at least test different es versions in the future.

action.auto_create_index=false does what it says. we don't want indices
to get created on the fly.

* Move utility methods from ElasticSearchBaseTest to custom client

* wrapped JestClient in custom client
* moved utility methods from ElasticSearchBaseTest to new client so that
  1.) the scope of ElasticSearchBaseTest gets narrowed down to providing
      the ES instance for the test suite
  2.) utility methods can be reused without having to inherit from base
      class
* cleaned up some tests

* Add license header

* Fixed type

* Bump testcontainers to latest version

* Use latest elasticsearch version as default in test container

* Reuse IndexMappingFactory in ElasticsearchBaseTest